Add files via upload
Rice variety classification is an important task in the agriculture and food industry. In this project, we
develop a Convolutional Neural Network (CNN) model to classify different varieties of rice based
on their images. The dataset consists of images of five rice varieties: Arborio, Basmati, Ipsala,
Jasmine, and Karacadag. We use the Keras library with TensorFlow backend for building and
training the CNN model. We experiment with different network architectures, activation
functions, optimizers, and hyperparameters to achieve optimal performance.